+++ lib/pages/driver.ex
@@ -14,35 +14,35 @@ defmodule Pages.Driver do
 
   @doc "Click an element within a page. Implementation for `Pages.click/4`."
   @callback click(Pages.Driver.t(), Pages.http_method(), Pages.text_filter() | nil, Hq.Css.selector()) ::
-              Pages.Driver.t() | no_return()
+              Pages.result() | no_return()
 
   @doc "Render a change. Implementation for `Pages.render_change/3`."
-  @callback render_change(Pages.Driver.t(), Hq.Css.selector(), Enum.t()) :: Pages.Driver.t()
+  @callback render_change(Pages.Driver.t(), Hq.Css.selector(), Enum.t()) :: Pages.result()
 
   @doc "Render a file upload. Implementation for `Pages.render_upload/4`."
-  @callback render_upload(Pages.Driver.t(), Pages.live_view_upload(), binary(), integer()) :: Pages.Driver.t()
+  @callback render_upload(Pages.Driver.t(), Pages.live_view_upload(), binary(), integer()) :: Pages.result()
 
   @doc "Render a hook event. Implementation for `Pages.render_hook/3`."
-  @callback render_hook(Pages.Driver.t(), binary(), Pages.attrs_t(), keyword()) :: Pages.Driver.t()
+  @callback render_hook(Pages.Driver.t(), binary(), Pages.attrs_t(), keyword()) :: Pages.result()
 
   @doc "Re-renders the page. Implementation for `Pages.rerender/1`."
-  @callback rerender(Pages.Driver.t()) :: Pages.Driver.t()
+  @callback rerender(Pages.Driver.t()) :: Pages.result()
 
   @doc "Submit a form targeted by the given selector. Implementation for `Pages.submit_form/2`."
-  @callback submit_form(Pages.Driver.t(), Hq.Css.selector()) :: Pages.Driver.t()
+  @callback submit_form(Pages.Driver.t(), Hq.Css.selector()) :: Pages.result()
 
   @doc "Fills in a form with the attributes and submits it. Implementation for `Pages.submit_form/4`."
-  @callback submit_form(Pages.Driver.t(), Hq.Css.selector(), atom(), Pages.attrs_t()) :: Pages.Driver.t()
+  @callback submit_form(Pages.Driver.t(), Hq.Css.selector(), atom(), Pages.attrs_t()) :: Pages.result()
 
   @doc "Fills in a form with the attributes and submits it. Implementation for `Pages.submit_form/5`."
   @callback submit_form(Pages.Driver.t(), Hq.Css.selector(), atom(), Pages.attrs_t(), Pages.attrs_t()) ::
-              Pages.Driver.t()
+              Pages.result()
 
   @doc "Fills in a form with the attributes without submitting it. Implementation for `Pages.update_form/4`."
-  @callback update_form(Pages.Driver.t(), Hq.Css.selector(), atom(), Pages.attrs_t()) :: Pages.Driver.t()
+  @callback update_form(Pages.Driver.t(), Hq.Css.selector(), atom(), Pages.attrs_t()) :: Pages.result()
 
   @doc "Navigate directly to a page. Implementation for `Pages.visit/2`."
-  @callback visit(Pages.Driver.t(), Path.t()) :: Pages.Driver.t()
+  @callback visit(Pages.Driver.t(), Path.t()) :: Pages.result()
 
   @doc "Target a child component for actions. Implementation for `Pages.with_child_component/3`."
   @callback with_child_component(Pages.Driver.t(), child_id :: binary(), (Pages.Driver.t() -> term())) ::